master <- Development

+++ SocialCareCaseViewerApi/V1/Factories/EntityFactory.cs
@@ -305,7 +305,7 @@ namespace SocialCareCaseViewerApi.V1.Factories
         {
             return new AllocationSet
             {
-                PersonId = request.MosaicId,
+                PersonId = (long) request.MosaicId,
                 WorkerId = workerId,
                 AllocationStartDate = allocationStartDate,
                 CaseStatus = caseStatus,
